% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{A sample example of a very good dataset.}
\format{a dataframe with two awesome columns and some really cool rows.
\describe{
  \item{Awesome}{awesomeness, in liters of blood, sweat, and tears}
  \item{Cool}{coolness, in kilocool points}
  ...
}}
\source{
I found it on the floor.
}
\usage{
exampleData
}
\description{
A tidy table of things
}
\keyword{datasets}
